trading_assistant/requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0

# Database
SQLAlchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# ML/AI
scikit-learn>=1.3.0
lightgbm>=4.0.0
xgboost>=2.0.0

# Embeddings & NLP
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# API & Web
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
pydantic>=2.0.0

# Scheduling
APScheduler>=3.10.0

# Email
jinja2>=3.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pyarrow>=13.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Market data
yfinance>=0.2.0

# Monitoring
psutil>=5.9.0
